Fixed the New Dimension operation.
The slot wasn't getting called for two reasons, it was looking for a
Cell type out of a copy-pasto, and NewDimOperation failed to call
QObject's init.

These menus are a good candidate for removing the sometimes overzealous
application of signals and slots in this codebase. The menus are
shortlived and have only one real user, therefore a callback would
probably be better.

Signed-off-by: Jon Kopetz <ccs4ever@gmail.com>